product controller update

+++ backend/src/controllers/productController.js
@@ -2,6 +2,11 @@ import Product from "../models/productModel.js";
 
 // Create a product
 export const createProduct = async (req, res) => {
+  if (!req.body.name || !req.body.sku || !req.body.createdBy || !req.body.categoryId || !req.body.supplierId) {
+    return res.status(400).json({ 
+      success: false,
+      message: "Name, SKU, CreatedBy, CategoryId, and SupplierId are required" });
+  }
   try {
     const ifExists = await product.findOne({ name: req.body.name });
     if (ifExists) {
@@ -25,10 +30,38 @@ export const getProducts = async (req, res) => {
   }
 };
 
+// Fetch products by category name
+export const getProductsByCategory = async (req, res) => {
+  try {
+    const { category } = req.params;
+    if (!category){
+      return res.status(400).json({
+        success: false,
+        message: "Category name is required" });
+    }
+
+    const products = await Product.find({ category });
+
+    res.status(200).json({
+      success: true,
+      count: products.length,
+      products,
+    });
+  } catch (error) {
+    console.error("Error fetching products by category:", error);
+    res.status(500).json({ success: false, message: "Server error" });
+  }
+};
+
 // Get a single product by ID
 export const getProduct = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id){
+      return res.status(400).json({
+        success: false,
+        message: "Product ID is required" });
+    }
     const product = await Product.findById(id);
 
     if (!product) {
@@ -45,6 +78,11 @@ export const getProduct = async (req, res) => {
 export const updateProduct = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id){
+      return res.status(400).json({
+        success: false,
+        message: "Product ID is required" });
+    }
     const product = await Product.findByIdAndUpdate(id, req.body, {
       new: true,
     });
@@ -63,6 +101,11 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id){
+      return res.status(400).json({
+        success: false,
+        message: "Product ID is required" });
+    }
     const product = await Product.findByIdAndDelete(id);
 
     if (!product) {
